requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
asyncpg==0.30.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Deprecated==1.2.14
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.4
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
multidict==6.1.0
netaddr==1.3.0
passlib==1.7.4
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-multipart==0.0.17
requests==2.32.3
sniffio==1.3.1
sqladmin==0.20.1
SQLAlchemy==2.0.36
starlette==0.41.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
wrapt==1.16.0
WTForms==3.1.2
yarl==1.17.1
yookassa==3.4.1
pytest==8.1.1